# Optimize toSearchParams