## Summary
I noticed toSearchParams showing up in profiles. I wrote a benchmarking script that revealed it had super-linear run times as the number of params grew: 

```
--- Benchmarking with 10 params ---
Original: 0.08ms per iteration
Optimized: 0.09ms per iteration

--- Benchmarking with 100 params ---
Original: 0.37ms per iteration
Optimized: 0.06ms per iteration

--- Benchmarking with 1000 params ---
Original: 6.62ms per iteration
Optimized: 0.72ms per iteration

--- Benchmarking with 10000 params ---
Original: 602.41ms per iteration
Optimized: 13.18ms per iteration

--- Benchmarking with 100000 params ---
Original: 50449.80ms per iteration
Optimized: 224.65ms per iteration
```

## Checklist
Delete items not relevant to your PR:
- [ ] Unit and integration tests covering the common scenarios were added
- [ ] A human-readable description of the changes was provided to include in CHANGELOG
- [ ] For significant changes, documentation in https://github.com/ClickHouse/clickhouse-docs was updated with further explanations or tutorials